introduce git-foreach script
This change introduces git-foreach -- a program that can be used to
perform one git operation on all git repositories that can be found by
searching the file system sub-tree starting with the current working
directory.
The script is meant to simplify working with large numbers of git
repositories. All that has to be done is to replace the 'git' command to
use in the normal case with the git-foreach script. For example, it is
possible to push the contents of all git repositories to the default
push destination, like so:
  $ git foreach push