do not show untracked files by default
When running git-foreach on a lot of repositories it is important to
keep the output concise. Some repositories have untracked files and
although these should eventually likely be ignored on the level of git
status itself (by means of gitignore), I believe it is also a reasonable
default to ignore them when performing git-foreach.
This change adds the --untracked-files=no option to the status
invocation to make untracked files disappear from the output.