add support for referencing the current git repository
This change introduces support for referencing the current git
repository on which we perform a command from within the command. This
mechanism is provided by means of the newly introduced REPO environment
variable that is defined for the very git command that is run within the
target repository.
An example invocation making use of this variable looks like this:
  $ git foreach remote add github git@github.com:d-e-s-o/'${REPO}'.git

Note: Because of shell expansion taking place when the git-foreach
      command is run (at least when done from a standard shell), the
      REPO variable must be escaped properly. In the example above this
      is achieved by putting its reference in single quotes.