refactor command execution code
An upcoming change will add dry-run support to allow the user to see in
advance what exactly is performed in each repository. This mode will
share parts of the command execution code. More precisely, the only
difference between this mode and actual execution is in the sed
replacement string.
This change factors out the replacement format command along with two
shared strings into newly introduced variables to allow for better code
sharing in the future.