introduce -n/--dry-run option
With the addition of REPO support it becomes more important to be able
to see what is about to being performed before actually running the git
command.
This change adds getopt-style command line argument parsing support and
introduces a new --dry-run option (short option: -n) that prints the
name of the repository and what exactly is being run in there.