introduce -h/--help options
With more options being added to the script it becomes more important to
be able to get an overview about them without having to look into the
source code.
To that end, this change introduces the --help option (short option: -h)
that displays a help text showing basic usage patterns as well as the
available options.